harvester/requirements.txt
# Geospatial
Fiona>=1.9.5,<1.9.999
geopandas>=0.14,<1.0
pyproj>=3.4,<3.7
Shapely>=2.0,<3.0

# Data processing
numpy>=1.23
pandas>=2.0

# Database
psycopg2-binary>=2.9

# AWS / Supabase storage
boto3>=1.28
botocore>=1.31

# HTTP & utilities
requests>=2.28
tqdm>=4.60
python-dotenv>=0.19
python-dateutil>=2.8
pytz>=2020.1

# CLI
click>=8.0
click-plugins>=1.1
cligj>=0.7

# Misc
attrs>=19.3
six>=1.14
munch>=2.5
